pyproject.toml
[project]
name = "freerag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.4.22",
    "gradio>=4.0.0",
    "huggingface-hub>=0.20.0",
    "ipykernel>=7.1.0",
    "llama-cpp-python>=0.2.50",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "rich>=13.7.0",
    "sentence-transformers>=2.2.2",
    "typer>=0.9.0",
]
